Fix code smell re maintainability - This file is part of one cycle involving 2 files
